feat(po): 단위 주문(kg/g) 1·2단계 — 공용 기반 + 판매자 등록폼 + RA 소수 발주
Fable 게이트 PASS · 마커 지문 a3738784979f
운영 미배포 — /배포 대기 (기존 4묶음 + SW bump 와 함께)

■ 1단계 — 전 역할 공용 기반 (화면 변화 0, 기본값=현행 동작)
- supplier_products.order_mode ENUM('pack','measure') 신설, DEFAULT 'pack'
  기존 38행 전부 pack 진입 검증. expandEnum 경유(목록 하드코딩 금지).
- min_order_quantity INT → DECIMAL(10,2) 2곳. 확폭 전후 건수·합계 불변 증명
  (sp 38행 86.00 / isp 58행 75.00). 멱등 재실행 EXIT 0.
  scripts/migrate-po-unit-order-foundation.js + registry deploy 등록.
- from-catalog 생성 흐름이 body 의 unit_conversion 을 무시하고 1 을 박던 버그 수정.
  이 하드코딩은 전 코드베이스 1곳뿐이었고 나머지 3벌은 원래 정상이었다.
  왕복 증명: 수정 전 1000 전송→1 저장 / 수정 후 1000 저장.
- min_order 정수 절삭 15곳을 공용 헬퍼로 일원화 (백엔드 utils/quantity.js,
  프론트 utils/unitConversion.ts). 그중 3곳은 Number.isInteger 검증으로
  0.5 를 400 으로 거부하고 있었다 — 판매자가 입력 자체를 못 하던 상태.
  왕복 증명: 0.5 → 1 (수정 전) / 0.5 → 0.5 (수정 후). 경계 10/10.
- R-SC-007 신설 — 단위 불일치 + 환산비 1 인 링크 감지(표시만, ⛔ 자동 백필 금지).
  seller_type 별 분리 조인(다형 참조 ID 충돌 방지) + collate 강제 + 개수 동의어 테이블
  + 표본 0 자가진단. 행 단위 검사라 baseline 이 구멍이 아니라 필터로 동작한다.
  반증: 동의어 무력화 5→8건 부활 / 조인 파괴 시 자가진단 발동 / 신규 주입 1건 EXIT 1 차단.

■ 2단계 — 판매자 등록폼 + RA 한정 구매 UI
- 공급업체 상품 등록에 주문 방식 라디오 2택(개수로 주문 / 무게·부피로 주문) + 1줄 미리보기.
  전문용어 없음, 기본=개수, 새 설정 페이지·토글·모달 0. 4언어.
- 발주 화면: measure 수량칸 단위 접미 + 소수 스텝, 규격 표시를 구조화 값 우선
  (이름에 박힌 "(900ml x 12btl)" 파싱은 폴백으로만), 단위당 가격으로 업체 비교.
- 최소주문 Math.max(전체 판매자) → 주문할 판매자 기준. 5곳 중 한 곳이 100 이면
  1개씩 파는 곳에서 사도 "Min 100" 이 뜨던 것 = 살 수 있는 물건을 못 사게 막던 표시.
- SupplierOptionModal 정수 강제 해제 — measure 만 소수, pack 은 정수 ±1 유지.
- 수량 표시 포매터 단일화(카트·staging·상세·인쇄). 로컬 fmtQty 제거.
- 기존 i18n 누락 newPo.minOrder 4언어 보충(변경 전부터 영어 fallback 이던 것).

■ 검증
- E2E 7/7: measure 등록(최소 0.5) → 재료 연결(conv 1000) → 1.5kg 발주 →
  1.4kg 입고 → 재고 정확히 +1400g. 잔재 0건.
- verify-all --full 16/16 (mount sweep 8역할 663.8s 크래시 0)
- print-guard 8/8 무접촉 · 고장 재현 반증 6회
- 실브라우저 UI 왕복: 화면 입력 → DB 도달 → 원복

■ 범위 밖(기록됨)
- 3단계 BG/FG measure UI — RA 실사용 후 게이트 제거만
- 서버 min_order 미강제 — 백로그 등재(사업 결정 선행, 재고 무결성 무관)
- order_mode/base_quantity 비스냅샷 — 확정 라인 표면이 현재값을 안 읽어 무영향(판정 기각)
- 환산비 4건 — 사람만 아는 값(계란 1판=몇 kg 등), Irene 회신 대기

+++ dev-backend/models/IngredientSellerProduct.js
@@ -24,7 +24,8 @@ const IngredientSellerProduct = sequelize.define('IngredientSellerProduct', {
     type: DataTypes.DECIMAL(10, 4), defaultValue: 1,
     comment: 'recipe unit -> seller product unit ratio'
   },
-  min_order_quantity: { type: DataTypes.INTEGER, defaultValue: 1 },
+  // measure 모드의 "최소 0.5kg" 을 담으려면 DECIMAL 이어야 한다 (INT 는 0.5 를 못 담음)
+  min_order_quantity: { type: DataTypes.DECIMAL(10, 2), defaultValue: 1 },
   lead_time_days: { type: DataTypes.INTEGER, defaultValue: 0 },
   is_preferred: { type: DataTypes.BOOLEAN, defaultValue: false },
   is_active: { type: DataTypes.BOOLEAN, defaultValue: true },
